pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "legal-mind"
version = "0.1.0"
description = "LegalMind Knowledge Assistant -- Modular RAG system for legal document Q&A"
requires-python = ">=3.11"
dependencies = [
    # LlamaIndex core + integrations
    "llama-index-core>=0.11.0",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-vector-stores-qdrant>=0.4.0",
    "llama-index-packs-node-parser-semantic-chunking>=0.4.0",
    "llama-index-readers-file>=0.4.0",
    # Vector store
    "qdrant-client[fastembed]>=1.12.0",
    # LLM & Embeddings
    "openai>=1.50.0",
    # Re-ranking
    "cohere>=5.20.0",
    # PDF parsing
    "pymupdf>=1.24.0",
    "pytesseract>=0.3.13",
    "Pillow>=10.4.0",
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # Cache
    "redis[hiredis]>=5.2.0",
    "numpy>=1.26.0",
    # Config
    "pydantic-settings>=2.6.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
    # Evaluation
    "deepeval>=2.0.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    # Utilities
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "reportlab>=4.4.10",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7.0",
    "mypy>=1.11.0",
    "ipython>=8.28.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
